feat(TransactionErrorCode) : Tx 도중 웹소켓 종료 시 오류코드 남기기
-웹소켓 종료 시점에 아직 Tx Ended 메시지를 받지 못한 TxInfo들을 찾아 오류코드를 01로 저장하고, 웹소켓 종료시점의 사용 정보를 토대로 금액, 사용량을 계산하여 저장한다.

+++ src/main/java/com/example/quaterback/websocket/transaction/event/converter/TransactionEventConverter.java
@@ -87,15 +87,15 @@ public class TransactionEventConverter {
     }
 
     public List<HourlyDischargeResponse> toHourlyDischargeResponseList(List<Object[]> raw) {
-        Map<Integer, Integer> hourMap = new HashMap<>();
+        Map<Integer, Double> hourMap = new HashMap<>();
         for (Object[] row : raw) {
             int hour = ((Number) row[0]).intValue();
-            int value = ((Number) row[1]).intValue();
+            double value = ((Number) row[1]).doubleValue();
             hourMap.put(hour, value);
         }
 
         return IntStream.range(0, 24)
-                .mapToObj(h -> new HourlyDischargeResponse(h, hourMap.getOrDefault(h, 0)))
+                .mapToObj(h -> new HourlyDischargeResponse(h, hourMap.getOrDefault(h, 0.0)))
                 .toList();
     }
